refactoring pinned message view to it's own file + optimizing colors for dark mode
